+++ CHANGELOG.md
@@ -3,5 +3,11 @@
 All notable changes to the "lobster" extension will be documented in this file.
 
 ## Releases
+- 0.0.4
+  - add confirm and cancel for dashboard
+- 0.0.3
+  - add some furniture and close up the demo counters
+- 0.0.2
+  - add some demo mcp counters
 - 0.0.1
   - initial release